Modificación eliminación de usuarios (se podía eliminar a sí mismo) + mensaje mostrando al usuario la falta de dinero si no tiene suficiente para realizar la compra.

+++ MyWallapop/src/main/java/com/uniovi/services/UsersService.java
@@ -43,12 +43,22 @@ public class UsersService {
 		return usersRepository.findByEmail(email);
 	}
 
-	public void deleteUser(ArrayList<Long> ids) {
-		for (Long id : ids)
-			usersRepository.deleteById(id);
+	public void deleteUser(ArrayList<Long> ids, User user) {
+		if(!ids.contains(user.getId()))
+		{
+			for (Long id : ids)
+				usersRepository.deleteById(id);
+		}
 	}
 
 	public void updateMoney(User user, Double price) {
 		usersRepository.updateMoney(user.getId(), price);
+	}
+	
+	public List<User> findUsersMinusPrincipal(User user)
+	{
+		List<User> userList =  this.getUsers();
+		userList.remove(user);
+		return userList;
 	}
 }
